Make shared base class for raw display
* Reduce amount of copy-paste
* Base class is abstract
  (too huge difference in concrete implementations)
* use `selectors.DefaultSelector` instead of `select.select`
  Automatically select OS specific optimal route by python stdlib.
